disk_block_cache: Create wrapper to store working set cache and sync cache.
